Greatly reduce bundle size. Zero dependencies!
Turns out we were only using three functions from ramda so it made sense
to use them on their own. After bringing curry, curryN and compose
in-house the bundled dropped from ~36kb to ~5kb. Ramda's implementation
of curry allows placeholders, which seems to be where the bloat comes
form.

The sizes listed above are completely unminified so even before this
commit it was a fairly small bundle. But now it's very small and
doesn't depend on anything else.